+++ apps/app/src/test/scala/org/lfdecentralizedtrust/splice/integration/tests/SvFrontendIntegrationTest.scala
@@ -480,6 +480,7 @@ class SvFrontendIntegrationTest
               element.underlying.sendKeys(voteReasonBody)
             }
             click on "save-vote-button"
+            click on "vote-confirmation-dialog-accept-button"
           },
         )(
           "sv2 can see the new vote request detail",
@@ -536,6 +537,7 @@ class SvFrontendIntegrationTest
               element.underlying.sendKeys(updatedVoteReasonBody)
             }
             click on "save-vote-button"
+            click on "vote-confirmation-dialog-accept-button"
           },
         )(
           "sv2 can see the new updated vote",
@@ -1305,6 +1307,7 @@ class SvFrontendIntegrationTest
                 click on "cast-vote-button"
                 click on "accept-vote-button"
                 click on "save-vote-button"
+                click on "vote-confirmation-dialog-accept-button"
               },
             )(
               "sv2's modal closes as all SVs voted before the expiration date",